MLE-23024 Cleaned up test plumbing
No longer need "AbstractJava17Test"
