+++ flux-cli/src/test/java/com/marklogic/flux/impl/importdata/ImportAndSplitFilesTest.java
@@ -5,15 +5,14 @@ package com.marklogic.flux.impl.importdata;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ArrayNode;
-import com.marklogic.flux.AbstractFluxTest;
-import com.marklogic.flux.AbstractJava17Test;
+import com.marklogic.flux.AbstractTest;
 import com.marklogic.junit5.XmlNode;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-class ImportAndSplitFilesTest extends AbstractJava17Test {
+class ImportAndSplitFilesTest extends AbstractTest {
 
     @Test
     void splitNamespacedXml() {
@@ -21,7 +20,7 @@ class ImportAndSplitFilesTest extends AbstractJava17Test {
             "import-files",
             "--path", "../flux-cli/src/test/resources/xml-file/namespaced-java-client-intro.xml",
             "--connection-string", makeConnectionString(),
-            "--permissions", AbstractFluxTest.DEFAULT_PERMISSIONS,
+            "--permissions", AbstractTest.DEFAULT_PERMISSIONS,
             "--collections", "files",
             "--uri-replace", ".*/xml-file,''",
             "--splitter-xpath", "/ex:root/ex:text/text()",
@@ -59,7 +58,7 @@ class ImportAndSplitFilesTest extends AbstractJava17Test {
             "import-files",
             "--path", "../flux-cli/src/test/resources/json-files/java-client-intro.json",
             "--connection-string", makeConnectionString(),
-            "--permissions", AbstractFluxTest.DEFAULT_PERMISSIONS,
+            "--permissions", AbstractTest.DEFAULT_PERMISSIONS,
             "--collections", "files",
             "--uri-replace", ".*/json-files,''",
             "--splitter-json-pointer", "/text",
@@ -83,7 +82,7 @@ class ImportAndSplitFilesTest extends AbstractJava17Test {
             "import-files",
             "--path", "../flux-cli/src/test/resources/json-files/java-client-intro.json",
             "--connection-string", makeConnectionString(),
-            "--permissions", AbstractFluxTest.DEFAULT_PERMISSIONS,
+            "--permissions", AbstractTest.DEFAULT_PERMISSIONS,
             "--collections", "files",
             "--uri-replace", ".*/json-files,''",
             "--splitter-json-pointer", ""
@@ -104,7 +103,7 @@ class ImportAndSplitFilesTest extends AbstractJava17Test {
             "import-files",
             "--path", "../flux-cli/src/test/resources/json-files/java-client-intro.json",
             "--connection-string", makeConnectionString(),
-            "--permissions", AbstractFluxTest.DEFAULT_PERMISSIONS,
+            "--permissions", AbstractTest.DEFAULT_PERMISSIONS,
             "--collections", "files",
             "--uri-replace", ".*/json-files,''",
             "--splitter-json-pointer", "/more-text",
@@ -127,7 +126,7 @@ class ImportAndSplitFilesTest extends AbstractJava17Test {
             "import-files",
             "--path", "../flux-cli/src/test/resources/json-files/java-client-intro.json",
             "--connection-string", makeConnectionString(),
-            "--permissions", AbstractFluxTest.DEFAULT_PERMISSIONS,
+            "--permissions", AbstractTest.DEFAULT_PERMISSIONS,
             "--uri-replace", ".*/json-files,''",
             "--splitter-json-pointer", "/text",
             "--splitter-custom-class", "com.marklogic.flux.impl.importdata.CustomSplitter",
@@ -149,7 +148,7 @@ class ImportAndSplitFilesTest extends AbstractJava17Test {
             "import-files",
             "--path", "../flux-cli/src/test/resources/json-files/java-client-intro.json",
             "--connection-string", makeConnectionString(),
-            "--permissions", AbstractFluxTest.DEFAULT_PERMISSIONS,
+            "--permissions", AbstractTest.DEFAULT_PERMISSIONS,
             "--splitter-json-pointer", "/text",
             "--splitter-custom-class", "doesnt.exist.ClassName"
         );
@@ -165,7 +164,7 @@ class ImportAndSplitFilesTest extends AbstractJava17Test {
             "import-files",
             "--path", "../flux-cli/src/test/resources/json-files/java-client-intro.json",
             "--connection-string", makeConnectionString(),
-            "--permissions", AbstractFluxTest.DEFAULT_PERMISSIONS,
+            "--permissions", AbstractTest.DEFAULT_PERMISSIONS,
             "--splitter-json-pointer", "/text",
             "--splitter-custom-class", "com.marklogic.flux.impl.importdata.splitter.CustomSplitter",
             "-Smissing an equals"
